feat: rm blockdata prior to Cancun and hub prior to Cancun
