+++ arithmetization/src/main/java/net/consensys/linea/zktracer/module/blockdata/module/BlockData.java
@@ -15,7 +15,9 @@
 package net.consensys.linea.zktracer.module.blockdata.module;
 
 import static net.consensys.linea.zktracer.Trace.LLARGE;
+import static net.consensys.linea.zktracer.TraceCancun.Blockdata.nROWS_DEPTH;
 import static net.consensys.linea.zktracer.module.ModuleName.BLOCK_DATA;
+import static net.consensys.linea.zktracer.module.blockdata.module.CancunBlockData.NB_ROWS_BLOCK_DATA;
 import static net.consensys.linea.zktracer.types.Conversions.bigIntegerToBytes;
 
 import java.util.*;
@@ -47,6 +49,9 @@ import org.hyperledger.besu.plugin.data.BlockHeader;
 
 @RequiredArgsConstructor
 public abstract class BlockData implements Module {
+
+  public static final short NB_ROWS_BLOCK_DATA = nROWS_DEPTH;
+
   private final Hub hub;
   private final Wcp wcp;
   private final Euc euc;
@@ -119,10 +124,8 @@ public abstract class BlockData implements Module {
   @Override
   public int lineCount() {
     final int numberOfBlock = blockInstructions.size() + (conflationFinished ? 0 : 1);
-    return numberOfBlock * numberOfLinesPerBlock();
+    return numberOfBlock * NB_ROWS_BLOCK_DATA;
   }
-
-  protected abstract int numberOfLinesPerBlock();
 
   @Override
   public int spillage(Trace trace) {
@@ -138,7 +141,6 @@ public abstract class BlockData implements Module {
   public void commit(Trace trace) {
 
     for (Map.Entry<Long, List<BlockDataInstruction>> entry : blockInstructions.entrySet()) {
-      Long key = entry.getKey();
       List<BlockDataInstruction> value = entry.getValue();
       for (BlockDataInstruction blockDataInstruction : value) {
         blockDataInstruction.trace(trace.blockdata());
