--- arithmetization/src/main/java/net/consensys/linea/zktracer/module/hub/ParisHub.java
@@ -1,38 +1,0 @@
-/*
- * Copyright ConsenSys Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package net.consensys.linea.zktracer.module.hub;
-
-import java.util.Map;
-
-import net.consensys.linea.zktracer.ChainConfig;
-import net.consensys.linea.zktracer.module.blockdata.module.BlockData;
-import net.consensys.linea.zktracer.module.blockdata.module.ParisBlockData;
-import net.consensys.linea.zktracer.module.euc.Euc;
-import net.consensys.linea.zktracer.module.wcp.Wcp;
-import net.consensys.linea.zktracer.types.PublicInputs;
-import org.apache.tuweni.bytes.Bytes;
-
-public class ParisHub extends LondonHub {
-  public ParisHub(ChainConfig chain, PublicInputs publicInputs) {
-    super(chain, publicInputs);
-  }
-
-  @Override
-  protected BlockData setBlockData(
-      Hub hub, Wcp wcp, Euc euc, ChainConfig chain, Map<Long, Bytes> blobBaseFees) {
-    return new ParisBlockData(hub, wcp, euc, chain, blobBaseFees);
-  }
-}
